requirements/development.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements/development.in -o requirements/development.txt
argon2-cffi==23.1.0
    # via -r requirements/base.in
argon2-cffi-bindings==25.1.0
    # via argon2-cffi
asgiref==3.11.1
    # via django
certifi==2026.1.4
    # via sentry-sdk
cffi==2.0.0
    # via argon2-cffi-bindings
cfgv==3.5.0
    # via pre-commit
coverage==7.13.3
    # via pytest-cov
distlib==0.4.0
    # via virtualenv
dj-database-url==2.3.0
    # via -r requirements/base.in
django==4.2.28
    # via
    #   -r requirements/base.in
    #   dj-database-url
    #   django-debug-toolbar
    #   djangorestframework
django-debug-toolbar==6.2.0
    # via -r requirements/development.in
djangorestframework==3.16.1
    # via -r requirements/base.in
factory-boy==3.3.3
    # via -r requirements/development.in
faker==40.1.2
    # via factory-boy
filelock==3.20.3
    # via virtualenv
gunicorn==22.0.0
    # via -r requirements/production.in
identify==2.6.16
    # via pre-commit
iniconfig==2.3.0
    # via pytest
nodeenv==1.10.0
    # via pre-commit
packaging==26.0
    # via
    #   gunicorn
    #   pytest
platformdirs==4.5.1
    # via virtualenv
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
pre-commit==4.5.1
    # via -r requirements/development.in
psycopg2-binary==2.9.11
    # via -r requirements/base.in
pycparser==3.0
    # via cffi
pygments==2.19.2
    # via pytest
pytest==9.0.2
    # via
    #   -r requirements/development.in
    #   pytest-cov
    #   pytest-django
pytest-cov==7.0.0
    # via -r requirements/development.in
pytest-django==4.11.1
    # via -r requirements/development.in
python-dotenv==1.2.1
    # via -r requirements/base.in
pyyaml==6.0.3
    # via pre-commit
ruff==0.15.0
    # via -r requirements/development.in
sentry-sdk==1.45.1
    # via -r requirements/production.in
sqlparse==0.5.5
    # via
    #   django
    #   django-debug-toolbar
typing-extensions==4.15.0
    # via dj-database-url
urllib3==2.6.3
    # via sentry-sdk
virtualenv==20.36.1
    # via pre-commit
whitenoise==6.11.0
    # via -r requirements/base.in
